requirements.txt
Django==4.2.2
moviepy==1.0.3
numpy==1.24.3
openai==0.27.8
opencv_python_headless==4.7.0.72
Pillow==9.0.1
python_docx==0.8.11
pytube==15.0.0
torch==2.0.1+cpu
vosk==0.3.45
